+++ duqtools/config/__init__.py
@@ -1,6 +1,3 @@
 from ._config import Config, cfg
-from ._runs import Run, Runs
-from ._system import AbstractSystem
-from ._workdir import WorkDirectory
 
-__all__ = ['cfg', 'WorkDirectory', 'Config', 'Runs', 'Run', 'AbstractSystem']
+__all__ = ['cfg', 'Config']
