--- duqtools/config/_description_helpers.py
@@ -1,7 +1,0 @@
-from textwrap import dedent
-
-
-def formatter(s):
-    """Dedent and remove newlines."""
-    s = dedent(s)
-    return s.replace('\n', ' ').strip()
