--- duqtools/config/basemodel.py
@@ -1,10 +1,0 @@
-from pydantic import BaseModel as PydanticBaseModel
-from pydantic import Extra
-from pydantic_yaml import YamlModelMixin
-
-
-class BaseModel(YamlModelMixin, PydanticBaseModel):
-    """Base model."""
-
-    class Config:
-        extra = Extra.forbid
