+++ duqtools/models/workdir.py
@@ -1,30 +1,13 @@
-from getpass import getuser
 from pathlib import Path
 from typing import List
 
-from pydantic import DirectoryPath, Field, validator
+from pydantic import validator
 
-from ._runs import Run, Runs
-from .basemodel import BaseModel
+from ..schema.runs import Run, Runs
+from ..schema.work_dir import WorkDirectoryModel
 
 
-class WorkDirectory(BaseModel):
-    """The workspace defines the the root directory where all simulations are
-    run. This is necessary, because some programs work with relative
-    directories from some root directory.
-
-    In this way, duqtools can ensure that the current work directory is
-    a subdirectory of the given root directory. All subdirectories are
-    calculated as relative to the root directory.
-
-    For example, for `rjettov`, the root directory is set to
-    `/pfs/work/$USER/jetto/runs/`. Any UQ runs must therefore be
-    a subdirectory.
-    """
-
-    root: DirectoryPath = Field(
-        f'/pfs/work/{getuser()}/jetto/runs/',
-        description='The directory from which experiments have to be run.')
+class WorkDirectory(WorkDirectoryModel):
 
     @property
     def cwd(self):
